changes on routes, user model created, register route created
